Tweak the cube styles & refactor
