+++ src/components/musicalCube/trackCubes/TrackCubes.css
@@ -1,0 +1,12 @@
+.one-shot-cube-container {
+    padding-top: 60px;
+}
+
+.melody-cube-container {
+    padding-top: 80px;
+}
+
+.only-melody-cube {
+    height: 100%;
+    padding-top: 200px;
+}
