+++ src/pages/realms/musicalCubes/MusicalCubesRealm.css
@@ -1,8 +1,19 @@
 .musical-realm { 
     height: 100%;
 
-    ion-row { 
+    ion-grid {
         height: 100%;
+    }
+
+    ion-row { 
         text-align: center;
+    }
+
+    .track-cubes-container {
+        height: 80%;
+    }
+
+    .track-picker-container {
+        height: 20%;
     }
 }
